#329: Splitted exaudflib cc into several files
