+++ exaudfclient/extension/rcontainer/rcontainer.h
@@ -1,7 +1,7 @@
 #ifndef RCONTAINER_H
 #define RCONTAINER_H
 
-#include "exaudflib/exaudflib.h"
+#include "exaudflib/vm/swig_vm.h"
   
 
 using namespace SWIGVMContainers;
